requirements.txt
# Loop Holonomy Deepfake Detection - Requirements
# Python 3.8+

# ============================================================================
# CORE DEPENDENCIES
# ============================================================================

# PyTorch (install with CUDA support if GPU available)
# Visit https://pytorch.org for installation command
torch>=2.0.0
torchvision>=0.15.0

# OpenAI CLIP (install separately)
# pip install git+https://github.com/openai/CLIP.git

# ============================================================================
# DATA & ML
# ============================================================================

numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pillow>=9.0.0

# ============================================================================
# VISUALIZATION
# ============================================================================

matplotlib>=3.5.0
seaborn>=0.12.0

# ============================================================================
# UTILITIES
# ============================================================================

tqdm>=4.64.0

# ============================================================================
# OPTIONAL: Experiment Tracking
# ============================================================================

# wandb>=0.15.0

# ============================================================================
# OPTIONAL: Advanced Visualization
# ============================================================================

# pacmap>=0.6.0          # For PaCMAP embeddings
# umap-learn>=0.5.0      # For UMAP embeddings

# ============================================================================
# DEVELOPMENT (for contributors)
# ============================================================================

# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.950
